refactor: share command palette contracts
